rename assetList.json
